% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YC.R
\name{YC}
\alias{YC}
\title{Compute the Yue and Clayton similarity between the rows of one or two matrices}
\usage{
YC(mat1, mat2 = NULL)
}
\arguments{
\item{mat1}{a \code{matrix} or \code{data.frame} where each row contains the data for a sample.}

\item{mat2}{(optional) a \code{matrix} or \code{data.frame} where each row contains the data for a sample.}
}
\value{
a similarity matrix.
If \code{mat2} is \code{NULL}, \code{YC} returns a square matrix of dimension
\code{n}, the number of rows of \code{mat1}.
Otherwise, \code{YC} returns a rectangular matrix with \code{n1 = nrow(mat1)} rows
and \code{n2 = nrow(mat2)} columns. Each element \code{[i,j]} gives the
Yue and Clayton similarity between \code{mat1[i,]} and \code{mat2[j,]}.
}
\description{
Compute the Yue and Clayton similarity between the rows of one or two matrices
}
